+++ src-tests/test-sha256.hs
@@ -4,8 +4,8 @@ module Main (main) where
 
 import           Data.ByteString        (ByteString)
 import qualified Data.ByteString        as B
-import qualified Data.ByteString.Lazy   as BL
 import qualified Data.ByteString.Base16 as B16
+import qualified Data.ByteString.Lazy   as BL
 
 -- reference implementation
 import qualified Data.Digest.Pure.SHA   as REF
@@ -124,6 +124,29 @@ rfc4231Tests = zipWith makeTest [1::Int ..] rfc4231Vectors
 
     hex = B16.encode
 
+rfc5869Vectors :: [(Int,ByteString,ByteString,ByteString,ByteString)]
+rfc5869Vectors = -- (l,ikm,salt,info,okm)
+    [ (42, rep 22 0x0b, x"000102030405060708090a0b0c", x"f0f1f2f3f4f5f6f7f8f9", x"3cb25f25faacd57a90434f64d0362f2a2d2d0a90cf1a5a4c5db02d56ecc4c5bf34007208d5b887185865")
+    , ( 82
+      , x"000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f202122232425262728292a2b2c2d2e2f303132333435363738393a3b3c3d3e3f404142434445464748494a4b4c4d4e4f"
+      , x"606162636465666768696a6b6c6d6e6f707172737475767778797a7b7c7d7e7f808182838485868788898a8b8c8d8e8f909192939495969798999a9b9c9d9e9fa0a1a2a3a4a5a6a7a8a9aaabacadaeaf"
+      , x"b0b1b2b3b4b5b6b7b8b9babbbcbdbebfc0c1c2c3c4c5c6c7c8c9cacbcccdcecfd0d1d2d3d4d5d6d7d8d9dadbdcdddedfe0e1e2e3e4e5e6e7e8e9eaebecedeeeff0f1f2f3f4f5f6f7f8f9fafbfcfdfeff"
+      , x"b11e398dc80327a1c8e7f78c596a49344f012eda2d4efad8a050cc4c19afa97c59045a99cac7827271cb41c65e590e09da3275600c2f09b8367793a9aca3db71cc30c58179ec3e87c14c01d5c1f3434f1d87"
+      )
+    , ( 42, rep 22 0x0b, "", "", x"8da4e775a563c18f715f802a063c5a31b8a11f5c5ee1879ec3454e5f3c738d2d9d201395faa4b61a96c8")
+    ]
+  where
+    x = fst.B16.decode
+    rep n c = B.replicate n c
+
+rfc5869Tests :: [TestTree]
+rfc5869Tests = zipWith makeTest [1::Int ..] rfc5869Vectors
+  where
+    makeTest i (l,ikm,salt,info,okm) = testGroup ("vec"++show i) $
+        [ testCase "hkdf" (hex okm @=? hex (IUT.hkdf l ikm salt info)) ]
+
+    hex = B16.encode
+
 -- define own 'foldl' here to avoid RULE rewriting to 'hashlazy'
 myfoldl' :: (b -> a -> b) -> b -> [a] -> b
 myfoldl' f z0 xs0 = lgo z0 xs0
@@ -191,5 +214,6 @@ main :: IO ()
 main = defaultMain $ testGroup "cryptohash-sha256"
     [ testGroup "KATs" katTests
     , testGroup "RFC4231" rfc4231Tests
+    , testGroup "RFC5869" rfc5869Tests
     , testGroup "REF" refImplTests
     ]